requirements.txt
pexpect==4.8.0
pyyaml